Implemented Call-Back mechanism.
The call-back mechanism is a global extension point which allows plugins
to register functions which can then be called from the WyccMain class.
In this case, the Whiley Build System plugin registers a main method
which basically takes care of all the actual compilation.

The purpose of this call back mechanism is to remove compilation
dependencies between the WhileyBuildSystem plugin and the
WhileyCompilerCollection project.  In essence, it means I can implement
the build system as a plugin, which is useful.