Fixed mistakes in AbstractConfigurable / ArrayUtils
Specifically, abstrsact configurable was only looking for methods
declared in the actual class, but not including those in super classes.
This was a problem when I tried to abstract commands in a common base
class for simplicity.

For ArrayUtils, the append() methods were all marked private which isn't
very helpful.