Remove AbstractConfigurable #3
This removes the concept of AbstractConfigurable.  This was a useful
piece of glue code, which roughly speaking replaced what was there
before.  There were too many problems with the way that it was working
via reflection.

Unfortunately, the overall structure of the command system remains
broken and needs to be overhauled.