Support build platforms attribute
This adds support for a build platforms attribute.  This attribute
is "build.platforms" and determines which build platforms a given
project is targetting.  For example, a project might be focused
primarily on JavaScript rather than native code, or Java bytecode, etc.

In additionan, this adds support for specifying default configuration values in schemas.
These defaults are used when no explicit value is given.