Support package installation #14
This adds true support for package installation.  More specifically, packages
are zip files containing WyIL files, wy.toml files, and other assets as
necessary (e.g. native code).

At this stage, a few things remains to be fixed.  Firstly, the manner in which
the package file is created is ugly as it breaks out of the wyfs and uses a
FileOutputStream; secondly, the wy.toml file is not being included which, of
course, is necessary.