Updates to Build Process
This pushes through a number of improvements to the way that binary files are
managed within the system.  In particular, we can now query the set of targets
along with the set of sources for those targets.  The concept of a Build.Graph
has been dropped.  Instead we have a Build.Executor which is responsible for
executing Build.Tasks in some appropriate order.  At the moment, the only
implementation is the SequentialBuildExecutor.  However, more will come.

One issue remaining is that a bug has been found in AbstractFolder.refresh().
Specifically, the problem is that refreshing after some new data has been
created essentially eliminates that data.  Therefore, we need a better refresh
mechanism.  In essence, this should only look for new files which should be
added to the list?