Add support for version resolution #53
This adds support for resolving the latest compatible version of any given
package.  For example, suppose 1.0.1 was requested and 1.1.5 existed.  Then, the
latter would always be chosen.

In order to implement this change, the central repository is given a master
index file which must be updated when pushing to the repository.

At this stage, version resolution is fairly simplistic and needs further
improvements.  For example, it will resolve the latest version of a package
even when the major version is 0 (which goes against semver.org).  Likewise, it
cannot gracefully handle multiple major versions of the same package (as this
would require true support for multi-version libraries).